tests/e2e/requirements.txt
pytest==7.4.3
pytest-html==4.1.1
psycopg2-binary==2.9.9
paho-mqtt==1.6.1
requests==2.31.0
